requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.1
celery==5.5.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
kombu==5.5.3
multidict==6.0.5
prompt_toolkit==3.0.51
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-weather==2.0.3
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.0
tzdata==2025.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
